feat: add LiteLLM routing, task queue, and proactive engine (#56)
* feat(llm): add LiteLLM multi-model routing provider

Add LiteLLMProvider as a 4th LLM provider, giving LabClaw access to
100+ models through one interface with fallback chains. Includes:
- LiteLLMProvider implementing LLMProvider protocol
- complete() and complete_structured() with JSON mode + Pydantic parse
- Fallback model support, timeout, retry, API key propagation
- 21 unit tests + 9 BDD scenarios, all passing


* feat(core): add background task queue with async execution

Add aiosqlite-backed task queue for background processing, enabling
experiment pipelines, retry logic, and scheduled analysis. Includes:
- TaskItem/ScheduleItem Pydantic schemas with priority and retry
- TaskQueue: enqueue, dequeue, status transitions, cancel
- TaskExecutor Protocol for safe, pluggable task execution
- TaskRunner: async polling loop with exponential backoff
- Event emission on all state transitions
- 30 unit tests + 10 BDD scenarios, all passing


* feat(core): add proactive engine for autonomous triggers and commitments

Add ProactiveEngine that bridges EventBus with TaskQueue for
autonomous 7x24 behavior. Includes:
- Trigger schema: event pattern matching (fnmatch), conditions, cooldown
- Commitment schema: tracking promises with overdue detection
- ProactiveEngine: event evaluation, task enqueue, lifecycle management
- Event emission: trigger_fired, commitment_overdue
- 36 unit tests + 10 BDD scenarios, all passing


* chore: add litellm dep, config sections, and coverage fix

- Add litellm>=1.30 to [llm] optional deps in pyproject.toml
- Add proactive/scheduling/litellm config sections to default.yaml
- Add daemon accumulator exception test for 100% coverage


* ci: add mypy typecheck job and ingest/nwb extras to CI

- Add typecheck job running mypy on src/labclaw/
- Add --extra ingest --extra nwb to test sync for full coverage


* fix(ci): resolve all CI failures — lint, mypy, bandit, test hang

- Format 6 files with ruff (task_queue, proactive, litellm tests/steps)
- Fix 21 pre-existing mypy errors across 14 files (type: ignore, annotations)
- Add # nosec B307 alongside # noqa: S307 for bandit eval() check
- Fix test suite hanging: close aiosqlite connections in BDD task_queue_steps
  (yield fixtures with cleanup prevent orphan threads blocking process exit)


* fix(core): catch executor exceptions in TaskRunner to prevent loop death

Wrap executor.execute() in try/except so an unhandled exception doesn't
kill the _run_loop asyncio task and leave tasks stuck in RUNNING forever.


---------

+++ src/labclaw/demo/generate.py
@@ -243,8 +243,8 @@ def _generate_chemistry() -> list[dict[str, Any]]:
     # -- Random exploration (remaining rows to reach 150) --
     remaining = 150 - len(rows)
     for _ in range(remaining):
-        temp = float(rng.integers(35, 100))
-        pres = float(rng.choice([100, 150, 200, 250, 300]))
+        temp = int(rng.integers(35, 100))
+        pres = int(rng.choice([100, 150, 200, 250, 300]))
         cat = float(rng.integers(5, 61))
         solvent = str(rng.choice(solvents))
